+++ packages/generator/src/edmx-to-vdm/v4/entity.spec.ts
@@ -1,22 +1,17 @@
-import { ServiceNameFormatter } from '../../../src/service-name-formatter';
-import { EdmxParameter, EdmxProperty } from '../../../src/edmx-parser/common';
 import {
   EdmxAction,
   EdmxActionImport,
   EdmxComplexType,
   EdmxEntitySet,
   EdmxEntityType
-} from '../../../src/edmx-parser/v4';
-import {
-  generateComplexTypesV4,
-  generateEntitiesV4
-} from '../../../src/edmx-to-vdm/v4';
-import { ServiceMetadata } from '../../../src/edmx-parser/edmx-file-reader';
-import { generateActionImportsV4 } from '../../../src/edmx-to-vdm/v4/action-import';
+} from '../../edmx-parser/v4';
+import { ServiceNameFormatter } from '../../service-name-formatter';
+import { EdmxProperty } from '../../edmx-parser/common';
+import { ServiceMetadata } from '../../edmx-parser/edmx-file-reader';
+import { generateComplexTypesV4 } from './complex-type';
+import { generateEntitiesV4 } from './entity';
 
-const defaultNamespace = 'ns';
-
-describe('edmx-to-vdm-v4', () => {
+describe('entity', () => {
   it('transforms collection type properties for primitive types', () => {
     const service = createTestServiceData(
       [
@@ -172,125 +167,67 @@ describe('edmx-to-vdm-v4', () => {
       fieldType: 'CollectionField'
     });
   });
-
-  it('transforms action imports', () => {
-    const formatter = getFormatter();
-    const service = createServiceWithActions();
-    const entites = generateEntitiesV4(service, [], [], formatter);
-    const actionImport = generateActionImportsV4(
-      service,
-      entites,
-      [],
-      formatter
-    );
-    expect(actionImport.length).toBe(2);
-
-    const actionNoReturnNoParameter = actionImport.find(
-      a => a.originalName === 'ActionNoReturnNoParameter'
-    );
-    expect(actionNoReturnNoParameter!.returnType.returnType).toBe('undefined');
-    expect(actionNoReturnNoParameter!.parameters).toStrictEqual([]);
-
-    const actionWithReturnWithParameter = actionImport.find(
-      a => a.originalName === 'ActionWithReturnWithParameter'
-    );
-    expect(actionWithReturnWithParameter!.returnType.returnType).toBe(
-      'TestEntity'
-    );
-    expect(actionWithReturnWithParameter!.parameters.length).toBe(1);
-  });
-
-  it('transforms entities from different namespaces', () => {
-    const entityType1 = createEntityType('TestEntityType1', [], [], 'ns1');
-    const entityType2 = createEntityType('TestEntityType2', [], [], 'ns2');
-    const entitySet1 = createTestEntitySet(
-      'TestEntity1',
-      'ns1.TestEntityType1',
-      [],
-      'ns3'
-    );
-    const entitySet2 = createTestEntitySet(
-      'TestEntity2',
-      'ns2.TestEntityType2',
-      [],
-      'ns3'
-    );
-    const complexType1 = createComplexType(
-      'TestComplexType1',
-      [createTestProperty('ComplexTypeProp')],
-      'ns1'
-    );
-    const complexType2 = createComplexType(
-      'TestComplexType2',
-      [createTestProperty('ComplexTypeProp')],
-      'ns2'
-    );
-
-    const actionReturnsEntityType = createAction(
-      'ActionReturnsEntityType',
-      'ns1.TestEntityType1',
-      [],
-      'n3'
-    );
-
-    const actionReturnsComplexType = createAction(
-      'ActionReturnsComplexType',
-      'ns2.TestComplexType2',
-      [],
-      'n3'
-    );
-
-    const service = createTestServiceData(
-      [entityType1, entityType2],
-      [entitySet1, entitySet2],
-      [complexType1, complexType2],
-      [actionReturnsEntityType, actionReturnsComplexType]
-    );
-
-    const complexTypes = generateComplexTypesV4(service, [], getFormatter());
-    expect(complexTypes.length).toBe(2);
-
-    const entities = generateEntitiesV4(
-      service,
-      complexTypes,
-      [],
-      getFormatter()
-    );
-    expect(entities.length).toBe(2);
-
-    const actionImports = generateActionImportsV4(
-      service,
-      entities,
-      complexTypes,
-      getFormatter()
-    );
-    expect(actionImports.length).toBe(2);
-  });
 });
 
-function getFormatter() {
+const defaultNamespace = 'ns';
+
+export function getFormatter() {
   return new ServiceNameFormatter();
 }
 
-function createServiceWithActions(): ServiceMetadata {
-  const entitySet = createTestEntitySet('TestEntity', 'TestEntityType', []);
-  const entityType = createEntityType('TestEntityType', [], []);
-  const actionNoReturnNoParameter = createAction(
-    'ActionNoReturnNoParameter',
-    undefined,
-    []
-  );
-  const actionWithReturnWithParameter = createAction(
-    'ActionWithReturnWithParameter',
-    'TestEntityType',
-    [{ Name: 'StringPara', Type: 'Edm.String' }]
-  );
-  const service = createTestServiceData([entityType], [entitySet], undefined, [
-    actionNoReturnNoParameter,
-    actionWithReturnWithParameter
-  ]);
+function createTestEntitySet(
+  name: string,
+  type: string,
+  navPropertyBindings: [string, string][] = [],
+  namespace: string = defaultNamespace
+): EdmxEntitySet {
+  return {
+    Namespace: namespace,
+    EntityType: `${type}`,
+    Name: name,
+    NavigationPropertyBinding: navPropertyBindings.map(([path, target]) => ({
+      Path: path,
+      Target: target
+    })),
+    'sap:content-version': '',
+    'sap:creatable': '',
+    'sap:deletable': '',
+    'sap:pageable': '',
+    'sap:updatable': ''
+  };
+}
 
-  return service;
+export function getComplexType(
+  namespace: string = defaultNamespace
+): EdmxComplexType {
+  return {
+    Name: 'TestComplexType',
+    Property: [createTestProperty('ComplexTypeProp', 'Edm.String')],
+    Namespace: namespace
+  };
+}
+
+function createTestProperty(name: string, type = 'Edm.String'): EdmxProperty {
+  return {
+    'sap:creatable': '',
+    'sap:filterable': '',
+    'sap:label': '',
+    'sap:sortable': '',
+    'sap:unicode': '',
+    'sap:updatable': '',
+    MaxLength: '10',
+    Nullable: '',
+    Name: name,
+    Type: type
+  };
+}
+
+function createImportsForActions(actions: EdmxAction[]): EdmxActionImport[] {
+  return actions.map(action => ({
+    Name: action.Name,
+    Action: `SomePrefix.${action.Name}`,
+    Namespace: action.Namespace
+  }));
 }
 
 function createTestServiceData(
@@ -324,54 +261,7 @@ function createTestServiceData(
   };
 }
 
-function getComplexType(namespace: string = defaultNamespace): EdmxComplexType {
-  return {
-    Name: 'TestComplexType',
-    Property: [createTestProperty('ComplexTypeProp', 'Edm.String')],
-    Namespace: namespace
-  };
-}
-
-function createComplexType(
-  name: string,
-  property: EdmxProperty[],
-  namespace: string = defaultNamespace
-): EdmxComplexType {
-  return {
-    Name: name,
-    Property: property,
-    Namespace: namespace
-  };
-}
-
-function createImportsForActions(actions: EdmxAction[]): EdmxActionImport[] {
-  return actions.map(action => ({
-    Name: action.Name,
-    Action: `SomePrefix.${action.Name}`,
-    Namespace: action.Namespace
-  }));
-}
-
-function createAction(
-  name: string,
-  returnType: string | undefined,
-  parameter: EdmxParameter[],
-  namespace: string = defaultNamespace
-): EdmxAction {
-  return {
-    Name: name,
-    Parameter: parameter,
-    ReturnType: returnType
-      ? {
-          Type: returnType
-        }
-      : undefined,
-    IsBound: false,
-    Namespace: namespace
-  };
-}
-
-function createEntityType(
+export function createEntityType(
   name: string,
   properties: [string, string, boolean][],
   navigationProperties: [string, string][] = [],
@@ -393,42 +283,5 @@ function createEntityType(
       Type: `namespace.${type}`
     })),
     Namespace: namespace
-  };
-}
-
-function createTestProperty(name: string, type = 'Edm.String'): EdmxProperty {
-  return {
-    'sap:creatable': '',
-    'sap:filterable': '',
-    'sap:label': '',
-    'sap:sortable': '',
-    'sap:unicode': '',
-    'sap:updatable': '',
-    MaxLength: '10',
-    Nullable: '',
-    Name: name,
-    Type: type
-  };
-}
-
-function createTestEntitySet(
-  name: string,
-  type: string,
-  navPropertyBindings: [string, string][] = [],
-  namespace: string = defaultNamespace
-): EdmxEntitySet {
-  return {
-    Namespace: namespace,
-    EntityType: `${type}`,
-    Name: name,
-    NavigationPropertyBinding: navPropertyBindings.map(([path, target]) => ({
-      Path: path,
-      Target: target
-    })),
-    'sap:content-version': '',
-    'sap:creatable': '',
-    'sap:deletable': '',
-    'sap:pageable': '',
-    'sap:updatable': ''
   };
 }
